Switch from accept4() to accept()
Commit ff2c503df091e6e4e9ab48cdb6df6ec8b7b525d0 introduced accept4() into
udev, which broke compatibility with kernels older than Linux 2.6.32 (or
Linux 2.6.36 on ARM). The purpose of accept4() is to permit
SOCK_NONBLOCK and O_CLOEXEC to be specified at the accept() call site
while previously, they would be specified either at the socket() call
site or throug a mix of the open() call site and a fcntl() call. The
existing socket() call already specifies SOCK_NONBLOCK and SOCK_CLOEXEC.
The Linux kernel documentation clearly states that accept() and
accept4() are equivalent when no flags are passed to accept4(), so
we are able to restore support for older kernels by replacing the
accept4() call with a call to accept().

Reported-by: Stephen Klimaszewski <steev@gentoo.org>
Signed-off-by: Richard Yao <ryao@gentoo.org>